svm: conformance: extract fd_hash into shared module
